Updates according to feedback: - Fixed typos, changed names and code style - Added `gte` & `lte` in filters and use the same for dates - Refactor the `GraphQLFactory` class to easily add new fields in the future
