+++ pom.xml
@@ -59,6 +59,7 @@
         <jackson.version>2.9.8</jackson.version>
         <log4j.version>2.11.2</log4j.version>
         <graphql.version>12.0</graphql.version>
+        <graphql.extended.scalars.version>1.0</graphql.extended.scalars.version>
     </properties>
 
     <licenses>
@@ -288,7 +289,7 @@
         <dependency>
             <groupId>com.graphql-java</groupId>
             <artifactId>graphql-java-extended-scalars</artifactId>
-            <version>1.0</version>
+            <version>${graphql.extended.scalars.version}</version>
         </dependency>
 
         <!-- Test dependencies -->
