+++ src/main/resources/crafter/engine/server-config.properties
@@ -229,7 +229,7 @@ crafter.engine.events.deployment.watcher.cron=0 * * ? * *
 # The path to look for content-type definitions
 crafter.engine.graphql.repoConfigFolder=/config/studio/content-types
 # The list of fields that should not be added to the GraphQL schema
-crafter.engine.graphql.ignoredFields=internal-name,placeInNav,orderDefault_f,folder-name
+crafter.engine.graphql.ignoredFields=internal-name,folder-name
 # The name of the file containing the content-type definition
 crafter.engine.graphql.contentType.definition=form-definition.xml
 # XPath selector for the content-type name
@@ -247,15 +247,15 @@ crafter.engine.graphql.contentType.xpath.property.title=title
 # XPath selector for the content-type property datasource
 crafter.engine.graphql.contentType.xpath.property.datasource.name=properties/property/name[text()='datasource']/../value
 # XPath selector for the content-type property item manager
-crafter.engine.graphql.contentType.xpath.property.itemmanager.name=properties/property/name[text()='itemManager']/../value
+crafter.engine.graphql.contentType.xpath.property.itemManager.name=properties/property/name[text()='itemManager']/../value
 # XPath selector for the content-type property child properties
 crafter.engine.graphql.contentTYpe.xpath.property.fields=fields/field
 # The name for the root Query type
-crafter.engine.graphql.query.name=Site
+crafter.engine.graphql.query.name=craftercms_site
 # The default value for the 'limit' argument
 crafter.engine.graphql.query.limit=10
 # The default value for the 'sortBy' argument
-crafter.engine.graphql.query.sort.field=createdDate_dt
+crafter.engine.graphql.query.sort.field=lastModifiedDate_dt
 # The default value for the 'sortOrder' argument
-crafter.engine.graphql.query.sort.order=ASC
+crafter.engine.graphql.query.sort.order=DESC
